refactor: Rename route handler function name to Usage
