+++ main.go
@@ -4,7 +4,6 @@ import (
 	"net/http"
 
 	"github.com/JanStanleyWatt/hukubiki-api/api"
-	"github.com/JanStanleyWatt/hukubiki-api/route"
 )
 
 func main() {
@@ -14,7 +13,7 @@ func main() {
 	}
 
 	http.HandleFunc("/int64n", api.Int64n)
-	http.HandleFunc("/", route.ResponseUsage)
+	http.HandleFunc("/", api.Usage)
 
 	server.ListenAndServe()
 }
